+++ plugins/rundown/app/components/RundownListItem/index.jsx
@@ -171,7 +171,8 @@ export function RundownListItem ({
     ]
 
     bridge.ui.contextMenu.open(spec, {
-      ...bridge.ui.contextMenu.getPositionFromEvent(e)
+      ...bridge.ui.contextMenu.getPositionFromEvent(e),
+      searchable: true
     })
   }