+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 ### Fixed
 - Select options now render in the system default color (#137)
 - Context menus now align themselvses with respect to the top and bottom of the viewport
+### Removed
+- Support for days in LTC
 
 ## 1.0.1
 ### Added