+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 ### Fixed
 - Select options now render in the system default color (#137)
 - Context menus now align themselvses with respect to the top and bottom of the viewport
+- An issue where the role popover had the wrong text size
 ### Removed
 - Support for days in LTC